added a start screen + enemy and player symbols + movement fixed + other fixes

+++ EnemyType1.h
@@ -7,7 +7,7 @@
 
 class EnemyType1 : public Enemy{
 public:
-    EnemyType1(int x, int y);
+    EnemyType1(double x, double y);
 };
